finish auth & users routes
