+++ README.md
@@ -12,8 +12,9 @@
 
 - Получает `email` и `password`
 - Делает валидацию с помощью Joi
-- В случае успешной валидации в модели User создает пользователя по данным которые прошли валидацию и отправляет письмо для верификации почты на введённую почту.
+- В случае успешной валидации в модели User создает пользователя по данным которые прошли валидацию возвращает `code:201`, обьект `data` с полями `id`, `email`, `avatar` и отправляет письмо для верификации почты на введённую почту.
 - Если почта уже используется кем-то другим, возвращает ошибку `Conflict`.
+- Возвращает обьект data с полями id, email, avatar.
 
 ### @ POST /auth/login
 
@@ -23,6 +24,7 @@
 - Если пароли совпадают и верификация пройдена: создает токен, сохраняет в текущем юзере и возвращает Успешный ответ.
 - Если пароль или имейл неверный, возвращает ошибку `'Email or password is wrong'`.
 - Если верификация почты не сделана, возвращает ошибку `'Please verify your email'`.
+- Если все проверки пройдены возвращает `code:200` и обьект `data` с полями `token` и `user`
 
 ### @ POST /auth/logout
 
@@ -44,11 +46,17 @@
 - Не получает body
 - Ищет информацию по токену из заголовка
 - Если пользователя не сущестует возвращает Ошибку `Unauthorized`.
-- В противном случае возвращает Успешный ответ
+- Если запрос успешно выполнен возвращает ответ вида:
+  "data": {
+  "user": {
+  "email": "test@test.com",
+  "avatarURL": "https://s.gravatar.com/avatar/b642b4217b34b1e8d3bd915fc65c4452?s=250"
+  }
+  }
 
 ### @ GET /users/verify/:verificationToken
 
-- Получает параметр verificationToken
+- Получает из строки запроса параметр verificationToken
 - Обрабатывает переход по ссылке верификации из письма
 - Сбрасывает verificationToken
 - При положительном результате возвращает ответ `Status: 200 OK, "Verification successful!"`
